+++ src/main/java/br/com/store/backend/infrastructure/rest/model/ErrorMessage.java
@@ -7,14 +7,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class ErrorMessage {
 
-    private Integer code;
+    private Long code;
     private String field;
     private String title;
     private String message;
 
     @JsonCreator
     public ErrorMessage(
-            @JsonProperty("code") Integer code, 
+            @JsonProperty("code") Long code, 
             @JsonProperty("field") String field,
             @JsonProperty("title") String title,
             @JsonProperty("message") String message) {
@@ -24,7 +24,7 @@ public class ErrorMessage {
         this.message = message;
     }
 
-    public Integer getCode() {
+    public Long getCode() {
         return code;
     }
